+++ Source/VagabondsWork/PlayerMainController.cpp
@@ -8,21 +8,13 @@
 #include "Ship.h"
 #include "Engine/LocalPlayer.h"
 #include "MapWidget.h"
+#include "PlayerSpectator.h"
 
 void APlayerMainController::BeginPlay()
 {
     Super::BeginPlay();
 
-    if (ShipMappingContext)
-    {
-        if (ULocalPlayer* LocalPlayer = GetLocalPlayer())
-        {
-            if (UEnhancedInputLocalPlayerSubsystem* Subsystem = LocalPlayer->GetSubsystem<UEnhancedInputLocalPlayerSubsystem>())
-            {
-                Subsystem->AddMappingContext(ShipMappingContext, 1);
-            }
-        }
-    }
+    UpdateShipMappingContext(GetPawn());
 
     if (MapWidgetClass && !MapWidgetInstance)
     {
@@ -30,6 +22,56 @@ void APlayerMainController::BeginPlay()
         if (MapWidgetInstance)
         {
             MapWidgetInstance->AddToViewport();
+        }
+    }
+}
+
+void APlayerMainController::OnPossess(APawn* InPawn)
+{
+    Super::OnPossess(InPawn);
+
+    if (APlayerSpectator* SpectatorPawn = Cast<APlayerSpectator>(InPawn))
+    {
+        CachedSpectatorPawn = SpectatorPawn;
+    }
+    UpdateShipMappingContext(InPawn);
+}
+
+void APlayerMainController::OnUnPossess()
+{
+    Super::OnUnPossess();
+
+    UpdateShipMappingContext(nullptr);
+}
+
+void APlayerMainController::UpdateShipMappingContext(APawn* InPawn)
+{
+    if (ULocalPlayer* LocalPlayer = GetLocalPlayer())
+    {
+        if (UEnhancedInputLocalPlayerSubsystem* Subsystem = LocalPlayer->GetSubsystem<UEnhancedInputLocalPlayerSubsystem>())
+        {
+            if (ShipMappingContext)
+            {
+                Subsystem->RemoveMappingContext(ShipMappingContext);
+            }
+            if (SpectatorMappingContext)
+            {
+                Subsystem->RemoveMappingContext(SpectatorMappingContext);
+            }
+            if (InPawn && InPawn->IsA<AShip>())
+            {
+                if (ShipMappingContext)
+                {
+                    Subsystem->AddMappingContext(ShipMappingContext, 1);
+                }
+            }
+            else if (InPawn && InPawn->IsA<APlayerSpectator>())
+            {
+                if (SpectatorMappingContext)
+                {
+                    Subsystem->AddMappingContext(SpectatorMappingContext, 0);
+                }
+            }
         }
     }
 }
@@ -64,6 +106,18 @@ void APlayerMainController::SetupInputComponent()
         EnhancedInputComponent->BindAction(RollAction, ETriggerEvent::Triggered, this, &APlayerMainController::HandleRollInput);
         EnhancedInputComponent->BindAction(RollAction, ETriggerEvent::Completed, this, &APlayerMainController::HandleRollInput);
     }
+    if (SpectatorMoveAction)
+    {
+        EnhancedInputComponent->BindAction(SpectatorMoveAction, ETriggerEvent::Triggered, this, &APlayerMainController::HandleSpectatorMoveInput);
+    }
+    if (SpectatorLookAction)
+    {
+        EnhancedInputComponent->BindAction(SpectatorLookAction, ETriggerEvent::Triggered, this, &APlayerMainController::HandleSpectatorLookInput);
+    }
+    if (SwitchControlAction)
+    {
+        EnhancedInputComponent->BindAction(SwitchControlAction, ETriggerEvent::Triggered, this, &APlayerMainController::HandleSwitchControlInput);
+    }
 }
 
 void APlayerMainController::HandleThrottleInput(const FInputActionValue& Value)
@@ -92,11 +146,67 @@ void APlayerMainController::HandleRollInput(const FInputActionValue& Value)
     UpdateShipRotationInput();
 }
 
+void APlayerMainController::HandleSpectatorMoveInput(const FInputActionValue& Value)
+{
+    if (APlayerSpectator* Spectator = Cast<APlayerSpectator>(GetPawn()))
+    {
+        Spectator->ApplyMoveInput(Value);
+    }
+}
+
+void APlayerMainController::HandleSpectatorLookInput(const FInputActionValue& Value)
+{
+    if (APlayerSpectator* Spectator = Cast<APlayerSpectator>(GetPawn()))
+    {
+        Spectator->ApplyLookInput(Value);
+    }
+}
+
+void APlayerMainController::HandleSwitchControlInput(const FInputActionValue& Value)
+{
+    if (Value.Get<bool>())
+    {
+        OnSwitchControlRequested();
+    }
+}
+
 void APlayerMainController::UpdateShipRotationInput()
 {
     if (AShip* Ship = Cast<AShip>(GetPawn()))
     {
         Ship->SetManualRotationInput(CachedPitchInput, CachedYawInput, CachedRollInput);
     }
+}
+
+void APlayerMainController::SwitchToPawn(APawn* PawnToControl)
+{
+    if (!PawnToControl)
+    {
+        return;
+    }
+
+    if (PawnToControl == GetPawn())
+    {
+        if (CachedSpectatorPawn.IsValid())
+        {
+            if (AShip* Ship = Cast<AShip>(GetPawn()))
+            {
+                CachedSpectatorPawn->SyncToTransform(Ship->GetShipCameraTransform());
+                CachedSpectatorPawn->SetCameraBoomLength(0.0f);
+            }
+            CachedSpectatorPawn->DetachFromActor(FDetachmentTransformRules::KeepWorldTransform);
+            Possess(CachedSpectatorPawn.Get());
+        }
+        return;
+    }
+
+    if (APlayerSpectator* SpectatorPawn = Cast<APlayerSpectator>(GetPawn()))
+    {
+        CachedSpectatorPawn = SpectatorPawn;
+        SpectatorPawn->SyncToTransform(PawnToControl->GetActorTransform());
+        SpectatorPawn->AttachToActor(PawnToControl, FAttachmentTransformRules::KeepWorldTransform);
+    }
+
+    Possess(PawnToControl);
 }